+++ binless/src/fast_dataframe.cpp
@@ -1,10 +1,10 @@
 #include "fast_dataframe.hpp"
-
+#include "graph_helpers.hpp"
 
 namespace binless {
 namespace fast {
 
-Rcpp::DataFrame get_as_dataframe(const FastData<Signal>& data, const DecayEstimator& dec) {
+Rcpp::DataFrame get_as_dataframe(const FastData<Signal>& data, const DecayEstimator& dec, double tol_val) {
   //bias, decay, signal with decay and exposures, and log_background matrix (w/ offset)
   std::vector<double> biasmat,decaymat,signal,binless,background;
   biasmat.reserve(data.get_N());
@@ -28,6 +28,17 @@ Rcpp::DataFrame get_as_dataframe(const FastData<Signal>& data, const DecayEstima
     binless.push_back(std::exp(ldec + lsig));
     background.push_back(nobs[i]*std::exp(bi + bj + ldec + exposure));
   }
+  //build patchnos
+  std::vector<double> patchnos;
+  patchnos.reserve(data.get_N());
+  for (unsigned dset=0; dset<data.get_ndatasets(); ++dset) {
+    std::vector<unsigned> d_bin1(dbin1.cbegin() + dset*data.get_ncells(), dbin1.cbegin() + (dset+1)*data.get_ncells());
+    std::vector<unsigned> d_bin2(dbin2.cbegin() + dset*data.get_ncells(), dbin2.cbegin() + (dset+1)*data.get_ncells());
+    std::vector<double> d_beta(log_signal.cbegin() + dset*data.get_ncells(), log_signal.cbegin() + (dset+1)*data.get_ncells());
+    IntegerVector patchno = get_patch_numbers(data.get_nbins(), tol_val, Rcpp::wrap(d_bin1), Rcpp::wrap(d_bin2), Rcpp::wrap(d_beta));
+    for (const double v : patchno) patchnos.push_back(v);
+  }
+  //return
   return Rcpp::DataFrame::create(_["name"]=data.get_name_factor(),
                                  _["bin1"]=data.get_bin1_factor(),
                                  _["bin2"]=data.get_bin2_factor(),
@@ -40,15 +51,27 @@ Rcpp::DataFrame get_as_dataframe(const FastData<Signal>& data, const DecayEstima
                                  _["signal"]=signal,
                                  _["binless"]=binless,
                                  _["phihat"]=data.get_signal_phihat(),
-                                 _["weight"]=data.get_signal_weights() );
+                                 _["weight"]=data.get_signal_weights(),
+                                 _["patchno"]=patchnos);
 }
 
 
-Rcpp::DataFrame get_as_dataframe(const FastData<Difference>& data) {
+Rcpp::DataFrame get_as_dataframe(const FastData<Difference>& data, double tol_val) {
   std::vector<double> log_difference = data.get_log_difference();
   std::vector<double> difference;
   difference.reserve(log_difference.size());
   for (const double v : log_difference) difference.push_back(std::exp(v));
+  //build patchnos
+  std::vector<unsigned> dbin1(data.get_bin1()), dbin2(data.get_bin2());
+  std::vector<double> patchnos;
+  patchnos.reserve(data.get_N());
+  for (unsigned dset=0; dset<data.get_ndatasets(); ++dset) {
+    std::vector<unsigned> d_bin1(dbin1.cbegin() + dset*data.get_ncells(), dbin1.cbegin() + (dset+1)*data.get_ncells());
+    std::vector<unsigned> d_bin2(dbin2.cbegin() + dset*data.get_ncells(), dbin2.cbegin() + (dset+1)*data.get_ncells());
+    std::vector<double> d_beta(log_difference.cbegin() + dset*data.get_ncells(), log_difference.cbegin() + (dset+1)*data.get_ncells());
+    IntegerVector patchno = get_patch_numbers(data.get_nbins(), tol_val, Rcpp::wrap(d_bin1), Rcpp::wrap(d_bin2), Rcpp::wrap(d_beta));
+    for (const double v : patchno) patchnos.push_back(v);
+  }
   return Rcpp::DataFrame::create(_["name"]=data.get_name_factor(),
                                  _["bin1"]=data.get_bin1_factor(),
                                  _["bin2"]=data.get_bin2_factor(),
@@ -57,7 +80,8 @@ Rcpp::DataFrame get_as_dataframe(const FastData<Difference>& data) {
                                  _["difference"]=difference,
                                  _["deltahat"]=data.get_deltahat(),
                                  _["weight"]=data.get_difference_weights(),
-                                 _["phi_ref"]=data.get_phi_ref());
+                                 _["phi_ref"]=data.get_phi_ref(),
+                                 _["patchno"]=patchnos);
 }
 
 }