+++ binless/src/fast_dataframe.hpp
@@ -12,9 +12,9 @@ using namespace Rcpp;
 namespace binless {
 namespace fast {
 
-Rcpp::DataFrame get_as_dataframe(const FastData<Signal>& data, const DecayEstimator& dec);
+Rcpp::DataFrame get_as_dataframe(const FastData<Signal>& data, const DecayEstimator& dec, double tol_val);
 
-Rcpp::DataFrame get_as_dataframe(const FastData<Difference>& data);
+Rcpp::DataFrame get_as_dataframe(const FastData<Difference>& data, double tol_val);
 
 }
 }