+++ CHANGELOG.md
@@ -8,6 +8,7 @@ for versions 0.x of binless, minor releases might break backwards compatibility.
 ### Added
 - First part of decay in fast binless is not forced to decrease. Use `free_diag`
  argument to decide (default 10kb)
+- patchno column to fast binless output
 
 ### Changed
 - Exposures are now fit separately